requirements.txt
Flask==2.3.2
Flask-Login==0.6.2
Flask-Cors==3.0.10
pymongo==4.5.0
flask-pymongo==2.3.0
python-dotenv==1.0.0
PyJWT==2.8.0
bcrypt==4.0.1
